Torrent upload megoldva

+++ src/main/java/hu/paalgyula/thetree/tags/ByteConveter.java
@@ -8,15 +8,15 @@ package hu.paalgyula.thetree.tags;
  * To change this template use File | Settings | File Templates.
  */
 public class ByteConveter {
-    public static String formatBytes( Long bytes ) {
+    public static String formatBytes(Long bytes) {
         String[] metrics = new String[]{"B", "KB", "MB", "GB", "TB"};
         int c = 0;
 
-        while ( ( bytes / 1024 ) > 1 ) {
+        while ((bytes / 1024) > 1) {
             c++;
-            bytes = bytes/1024;
+            bytes = bytes / 1024;
         }
 
-        return (double)(Math.ceil( bytes * 100 ) / 100) + " " + metrics[c];
+        return Math.ceil(bytes * 100) / 100 + " " + metrics[c];
     }
 }
